Adicão de modificacoes para openMP e MPI

+++ projeto_final/City.cpp
@@ -3,7 +3,7 @@
 #include "City.h"
 
 
-City::City(float x, float y, int id){
+City::City(double x, double y, int id){
     this->x=x;
     this->y=y;
     this->id=id;
@@ -13,8 +13,8 @@ void City::print(){
     printf("City(%f, %f, %d)\n", x, y, id);
 }
 
-float City::distance(City* otherCity){
-    float dx = x - (otherCity->x);
-    float dy = y - (otherCity->y);
+double City::distance(City* otherCity){
+    double dx = x - (otherCity->x);
+    double dy = y - (otherCity->y);
     return sqrt( dx*dx + dy*dy );
 }
